New version released to Fx Marketplace

+++ Makefile
@@ -1,5 +1,5 @@
 zip:
-	rm authenticator.zip
+	if [ -f authenticator.zip ]; then rm authenticator.zip; fi
 	zip -r authenticator.zip \
 		bower_components/building-blocks/style/buttons.css \
 		bower_components/building-blocks/style/headers.css \
@@ -10,6 +10,11 @@ zip:
 		bower_components/building-blocks/style/value_selector.css \
 		bower_components/building-blocks/util.css \
 		bower_components/building-blocks/fonts.css \
+		bower_components/building-blocks/fonts/FiraSans/FiraSans-Regular.woff \
+		bower_components/building-blocks/fonts/FiraSans/FiraSans-Light.woff \
+		bower_components/building-blocks/style/headers/images/icons/add.png \
+		bower_components/building-blocks/style/headers/images/icons/back.png \
+		bower_components/building-blocks/style/headers/images/icons/close.png \
 		style/ \
 		bower_components/jquery/dist/jquery.min.js \
 		bower_components/handlebars/handlebars.min.js \
